Extract agent final answer builder
