+++ docs/DECISIONS.md
@@ -209,3 +209,4 @@
 - Composer submission planning lives in `WorkspaceComposerSubmissionPlanner`. The planner owns draft trimming plus the first routing decision: ignore blank drafts, dispatch slash commands with their trimmed original prompt, or send a trimmed agent prompt. `WorkspaceModel` owns the resulting side effects, not raw prompt normalization.
 - Transcript timeline layout lives in `QuillCodeTranscriptView`, not the workspace shell. The workspace shell decides whether the transcript pane is visible and routes callbacks; the transcript view owns Find bar placement, empty state, context banner/runtime issue/review placement, message and tool-card timeline rows, active Find highlighting, scroll-to-match behavior, and tool-card copy fallback. `WorkspaceSwiftUIView.swift` should stay focused on chrome composition, modal presentation, and typed command routing.
 - Workspace sheet presentation lives in `QuillCodeWorkspaceSheetsModifier`, not the workspace shell. `WorkspaceSwiftUIView.swift` owns the presentation state bindings and routes typed callbacks; the sheet presenter owns settings, search, keyboard shortcuts, command palette, worktree create/remove, and thread/project rename sheet composition plus their save/cancel dismissal wiring.
+- Tool-result final-answer copy lives in `AgentFinalAnswerBuilder`, not in `AgentRunner`. `AgentRunner.finalAnswer(...)` remains the compatibility entry point, but shell/browser/file/patch/MCP/Computer Use response wording belongs in the focused builder with direct tests and parity gates.
